Remove platforms that aren't supported by GlusterFS or Docker
GlusterFS doesn't support Windows clients so it we don't need to
create binaries for Windows. Darwin isn't directly supported by
Docker, though it seems likely to be coming soon. (After all, it
is a *nix like OS.) However, it's unlikely that anyone is going
to want to run the 386 version of this project. Therefore, I
removed the 386 compile and left in support for the x64 platform.